week08/makaoBank/220809-조성환/bank/src/test/java/com/makao/bank/repositories/AccountRepositoryTest.java
package com.makao.bank.repositories;

import com.makao.bank.models.Account;
import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;

class AccountRepositoryTest {
  @Test
  void find() {
    AccountRepository accountRepository = new AccountRepository();

    Account account = accountRepository.find("1234");

    assertThat("1234").isEqualTo(account.identifier());
  }

  @Test
  void notFound() {
    AccountRepository accountRepository = new AccountRepository();

    Account account = accountRepository.find("-1235");

    assertThat(account).isNull();
  }
}
